feat(auto-refresh): add a token expiry check to isAuthenticatedFactory
Prevents edge cases where users accessing `getKindeServerSession` on an RSC that isn't protected by their middleware, after their tokens have expired, will be operating with stale tokens.

If the token is expired here, we'll redirect to login.

Note: this would probably be better in the `getKindeServerSession` root, but we'd need to make it async. Maybe something to consider down the track.