refactor: use new redirectOnExpiredToken function
Sidenote: This technically only prevents the edge cases in RSCs for getAccessToken, getIdToken and isAuthenticated. We could add this to every server session factory if need be.

Ultimately the real fix would be to have this at the root of getKindeServerSession, but that would require us to make getKindeServerSession async, which is breaking.