fix: rework prefetch check
The current logic assumes headers is a map (which is true for Next >14), but in Next 13 the request headers are an object. This causes errors on Next 13 like so: TypeError: e.headers.get is not a function

Given the headers function was added in 13, we should resort to using that instead as it provides a stable and consistent API for accessing headers in all versions greater than Next 13.